comments out code in the reducers and actions files in the Groups comoponent folder and creates them in to two separate folders in a /redux directory
