+++ src/login/components/App/DashboardApp/Settings/Groups/Groups_actions.js
@@ -1,19 +1,19 @@
-// export const POST_USERDATA = "POST_USERDATA";
-export const GET_GROUP_MANAGEMENT_GROUPS_SUCCESS = "GET_GROUP_MANAGEMENT_GROUPS_SUCCESS";
-export const GET_GROUP_MANAGEMENT_GROUPS_FAIL = "GET_GROUP_MANAGEMENT_GROUPS_FAIL";
+// // export const POST_USERDATA = "POST_USERDATA";
+// export const GET_GROUP_MANAGEMENT_GROUPS_SUCCESS = "GET_GROUP_MANAGEMENT_GROUPS_SUCCESS";
+// export const GET_GROUP_MANAGEMENT_GROUPS_FAIL = "GET_GROUP_MANAGEMENT_GROUPS_FAIL";
 
-export function getAllUserdata() {
-  return {
-    type: GET_ALL_USERDATA
-  };
-}
+// export function getAllUserdata() {
+//   return {
+//     type: GET_ALL_USERDATA
+//   };
+// }
 
-export function getGropusDataFail(err) {
-  return {
-    type: GET_GROUP_MANAGEMENT_GROUPS_FAIL,
-    error: true,
-    payload: {
-      message: err
-    }
-  };
-}
+// export function getGropusDataFail(err) {
+//   return {
+//     type: GET_GROUP_MANAGEMENT_GROUPS_FAIL,
+//     error: true,
+//     payload: {
+//       message: err
+//     }
+//   };
+// }
