+++ src/login/components/App/DashboardApp/Settings/Groups/Groups_reducer.js
@@ -1,26 +1,26 @@
-import * as actions from "./Groups_actions";
+// import * as actions from "./Groups_actions";
 
-const groupsData = {
-  message: "",
-  data: {
-    // eppn: "",
-    // given_name: "",
-    // surname: "",
-    // display_name: "",
-    // language: "",
-  },
-};
+// const groupsData = {
+//   message: "",
+//   data: {
+//     // eppn: "",
+//     // given_name: "",
+//     // surname: "",
+//     // display_name: "",
+//     // language: "",
+//   },
+// };
 
-let groupsDataReducer = (state = groupsData, action) => {
-  switch (action.type) {
-    case actions.GET_GROUP_MANAGEMENT_GROUPS_SUCCESS:
-      return {
-        ...state,
-        data: { ...action.payload },
-      };
-    default:
-      return state;
-  }
-};
+// let groupsDataReducer = (state = groupsData, action) => {
+//   switch (action.type) {
+//     case actions.GET_GROUP_MANAGEMENT_GROUPS_SUCCESS:
+//       return {
+//         ...state,
+//         data: { ...action.payload },
+//       };
+//     default:
+//       return state;
+//   }
+// };
 
-export default groupsDataReducer;
+// export default groupsDataReducer;
