+++ src/login/redux/reducers/getAllGroupsData_reducer.js
@@ -1,0 +1,20 @@
+import * as actions from "../actions/getAllGroupsData_actions";
+
+const groupsData = {
+  message: "",
+  data: {},
+};
+
+let groupsDataReducer = (state = groupsData, action) => {
+  switch (action.type) {
+    case actions.GET_GROUP_MANAGEMENT_ALL_DATA_SUCCESS:
+      return {
+        ...state,
+        data: { ...action.payload },
+      };
+    default:
+      return state;
+  }
+};
+
+export default groupsDataReducer;
